+++ apps/recommendation_app/src/main/java/org/sab/recommendation/GetThreadRecommendationCommand.java
@@ -29,7 +29,7 @@ public class GetThreadRecommendationCommand {
                 thread.setName(o.getString("name"));
                 thread.setDescription(o.getString("description"));
                 thread.setCreator(o.getString("creator"));
-                thread.setNumOfFollowers(o.getInt("numOfFollowers"));
+                thread.setNumOfFollowers(o.getLong("numOfFollowers"));
                 thread.setDateCreated((String) o.get("dateCreated"));
                 System.out.println("Recommendation Results " + thread);
             }
